using former RSS State Machine as Core StateMachine
